Separate types module
