--- aionetworking/formats/contrib/types.py
@@ -1,8 +1,0 @@
-from __future__ import annotations
-from typing import TYPE_CHECKING, TypeVar
-
-if TYPE_CHECKING:
-    from .json import JSONObject
-
-
-JSONObjectType = TypeVar('JSONObjectType', bound='JSONObject')
