--- aionetworking/types.py
@@ -1,131 +1,0 @@
-from __future__ import annotations
-from collections import ChainMap
-import builtins
-from dataclasses import dataclass
-import operator
-
-from ipaddress import IPv4Network, IPv6Network, AddressValueError
-
-from typing import Union, Callable, Iterable, Any, Sequence
-
-
-
-
-class IPNetwork:
-    def __init__(self, network: Union[str, IPv4Network, IPv6Network]):
-        if isinstance(network, IPv4Network):
-            self.ip_network = network
-            self.is_ipv6 = False
-        elif isinstance(network, IPv6Network):
-            self.ip_network = network
-            self.is_ipv6 = True
-        else:
-            try:
-                self.ip_network = IPv4Network(network)
-                self.is_ipv6 = False
-            except AddressValueError:
-                try:
-                    self.ip_network = IPv6Network(network)
-                    self.is_ipv6 = True
-                except AddressValueError:
-                    raise AddressValueError(f'{network} is not valid IPv4 or IPv6 network')
-
-    def __eq__(self, other):
-        return self.ip_network == other.ip_network
-
-    def supernet_of(self, network: Union[IPv4Network, IPv6Network]):
-        return self.ip_network.supernet_of(network)
-
-
-def supernet_of(network: Union[str, IPNetwork, IPv4Network, IPv6Network], networks: Sequence[IPNetwork]):
-    if not isinstance(network, IPNetwork):
-        network = IPNetwork(network)
-    if network.is_ipv6:
-        networks = filter(lambda n: n.is_ipv6, networks)
-    else:
-        networks = filter(lambda n: not n.is_ipv6, networks)
-    return any(n.supernet_of(network.ip_network) for n in networks)
-
-
-@dataclass
-class CallableFromString(Callable):
-    _strings_to_callables = {}
-    callable: Callable
-
-    def __post_init__(self):
-        self.callable = self.adapt_callable(self.callable)
-
-    def __call__(self, *args, **kwargs):
-        return self.callable(*args, **kwargs)
-
-    def adapt_callable(self, v: Union[str, Callable]):
-        if isinstance(v, str):
-            try:
-                return self._strings_to_callables[v]
-            except KeyError:
-                raise TypeError(f"{v} not found")
-        if v not in self._strings_to_callables.values():
-            name = self.__class__.__name__
-            raise TypeError(f"{v} not a valid {name}")
-        return v
-
-
-class Builtin(CallableFromString):
-    _strings_to_callables = ChainMap(builtins.__dict__, {'istr': str})
-
-
-def in_(a, b) -> bool:
-    return a in b
-
-
-class Operator(CallableFromString):
-    _strings_to_callables = {
-        '=': operator.eq,
-        '==': operator.eq,
-        'eq': operator.eq,
-        '<': operator.lt,
-        'lt': operator.lt,
-        '<=': operator.le,
-        'lte': operator.le,
-        '!<': operator.ne,
-        'ne': operator.ne,
-        '>': operator.gt,
-        'gt': operator.gt,
-        '>=': operator.ge,
-        'ge': operator.ge,
-        'in': in_,
-        'contains': operator.contains
-    }
-
-
-@dataclass
-class Expression:
-    attr: str
-    op: Operator
-    value_type: Builtin
-    value: Any
-    case_sensitive: bool = True
-
-    @classmethod
-    def from_string(cls, string: str):
-        attr, op, value_type, value = string.split()
-        if op.startswith('i'):
-            case_sensitive = True
-            op = op.split('i')[1]
-        else:
-            case_sensitive = False
-        op = Operator(op)
-        return cls(attr, op, value_type, value, case_sensitive=case_sensitive)
-
-    def __call__(self, obj: Any) -> bool:
-        if not self.attr or self.attr == 'self':
-            value = obj
-        else:
-            value = getattr(obj, self.attr)
-        if self.case_sensitive:
-            if isinstance(value, Iterable):
-                value = [v.lower() for v in value]
-            else:
-                value = value.lower()
-            return self.op(value, self.value.lower())
-        return self.op(value, self.value)
