+++ aionetworking/types/conf.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, TypeVar
 
 if TYPE_CHECKING:
-    from .logging import BaseLogger, ConnectionLogger
+    from aionetworking.conf.logging import BaseLogger, ConnectionLogger
 
 
 LoggerType = TypeVar('LoggerType', bound='BaseLogger')
