+++ aionetworking/types/receivers.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, TypeVar
 
 if TYPE_CHECKING:
-    from .protocols import ReceiverProtocol
+    from aionetworking.receivers.protocols import ReceiverProtocol
 
 
 ReceiverType = TypeVar('ReceiverType', bound='ReceiverProtocol')
