+++ aionetworking/types/senders.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, TypeVar
 
 if TYPE_CHECKING:
-    from .protocols import SenderProtocol
+    from aionetworking.senders.protocols import SenderProtocol
 
 
 SenderType = TypeVar('SenderType', bound='SenderProtocol')
