+++ tests/test_config/conftest.py
@@ -3,7 +3,7 @@ import yaml
 from aionetworking.conf.log_filters import PeerFilter, MessageFilter
 from aionetworking.conf.logging import Logger, ConnectionLogger, ConnectionLoggerStats, StatsTracker, StatsLogger
 from aionetworking.conf.yaml_config import load_all_tags, get_paths
-from aionetworking.types import Expression
+from aionetworking.utils import Expression
 from tests.conftest import get_fixture
 from tests.test_senders.conftest import *
