+++ tests/test_end_to_end/test_in_executors.py
@@ -5,8 +5,7 @@ from pathlib import Path
 import os
 
 ###Required for skipif in fixture params###
-from aionetworking.compatibility import datagram_supported
-from aionetworking.utils import supports_pipe_or_unix_connections
+from aionetworking.compatibility import datagram_supported, supports_pipe_or_unix_connections
 
 
 class TestOneWaySenderInExecutors:
