+++ tests/test_end_to_end/test_single_loop.py
@@ -7,8 +7,7 @@ from aionetworking.utils import alist
 from aionetworking.actions.echo import InvalidRequestError
 
 ###Required for skipif in fixture params###
-from aionetworking.compatibility import datagram_supported
-from aionetworking.utils import supports_pipe_or_unix_connections
+from aionetworking.compatibility import datagram_supported, supports_pipe_or_unix_connections
 
 
 class TestOneWayServer:
