+++ tests/test_formats/conftest.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from aionetworking.formats.contrib.json import JSONObject, JSONCodec
 from aionetworking.formats.recording import BufferCodec, BufferObject, recorded_packet
-from aionetworking.formats.types import MessageObjectType
+from aionetworking.types.formats import MessageObjectType
 
 from typing import Dict, Any, List, NamedTuple, Tuple, Type
