+++ tests/test_networking/test_adaptors_one_way.py
@@ -2,7 +2,7 @@ import pytest
 import asyncio
 from pathlib import Path
 
-from aionetworking.formats.recording import get_recording_from_file
+from aionetworking.formats import get_recording_from_file
 from aionetworking.utils import alist, time_coro
