+++ tests/test_networking/test_adaptors_two_way.py
@@ -3,7 +3,7 @@ import asyncio
 
 from pathlib import Path
 
-from aionetworking.formats.recording import get_recording_from_file
+from aionetworking.formats import get_recording_from_file
 from aionetworking.networking.exceptions import MethodNotFoundError
 from aionetworking.utils import alist
