+++ tests/test_networking/test_sftp.py
@@ -3,7 +3,7 @@ import pytest
 import pickle
 from pathlib import Path
 
-from aionetworking.formats.recording import get_recording_from_file
+from aionetworking.formats import get_recording_from_file
 from aionetworking.utils import alist
